requirements.txt
numpy
matplotlib
pandas
pyserial
python-vxi11
vector
scipy
